refactor(trading): move getRandomAccountDescriptor to suite-common
Extract getRandomAccountDescriptor into @suite-common/trading as the single
source and have both web (TRADING_FALLBACK_API_KEY) and native consume it,
replacing the duplicated native-only helper.
