+++ suite-common/trading/src/utils/apiKeyUtils.ts
@@ -1,0 +1,3 @@
+import { getWeakRandomId } from '@trezor/utils';
+
+export const getRandomAccountDescriptor = () => getWeakRandomId(20);
