+++ suite-native/module-trading/src/hooks/buy/useBuyData.ts
@@ -1,8 +1,11 @@
 import { useEffect } from 'react';
 import { useDispatch, useSelector } from 'react-redux';
 
-import { selectTradingBuyLoadingTimestampAndStatus, tradingThunks } from '@suite-common/trading';
-import { getRandomAccountDescriptor } from '@suite-native/trading-quote-utils';
+import {
+    getRandomAccountDescriptor,
+    selectTradingBuyLoadingTimestampAndStatus,
+    tradingThunks,
+} from '@suite-common/trading';
 import { selectBuySelectedReceiveAccount } from '@suite-native/trading-state';
 
 export const useBuyData = (reloadRequestOrdinal: number) => {
