+++ suite-native/module-trading/src/hooks/exchange/useExchangeData.ts
@@ -2,10 +2,10 @@ import { useEffect } from 'react';
 import { useDispatch, useSelector } from 'react-redux';
 
 import {
+    getRandomAccountDescriptor,
     selectTradingExchangeLoadingTimestampAndStatus,
     tradingThunks,
 } from '@suite-common/trading';
-import { getRandomAccountDescriptor } from '@suite-native/trading-quote-utils';
 import { selectExchangeSelectedSendAccount } from '@suite-native/trading-state';
 
 export const useExchangeData = (reloadRequestOrdinal: number) => {
